Fixed all parking issues
NOTE: There are poissibilities the vehicle parks the next day but that is fine.

+++ src/ActivityEngine.h
@@ -108,8 +108,8 @@ private:
     void process_departure_events(SimTime &sim_time, VehicleType &veh_type, VehicleStats *veh);
     void simulate_events();
     long double biased_expovariate(double rate_param, double lower_bound, double upper_bound);
-    long double estimate_departure_time(VehicleStats &veh, simtime_t start_timestamp);
-    long double estimate_departure_delta(VehicleStats &veh, double speed);
+    double estimate_departure_time(VehicleStats &veh, simtime_t start_timestamp);
+    double estimate_departure_delta(VehicleStats &veh, double speed);
     mt19937_64 mersenne_twister_engine;
     unsigned n_vehicles_monitored, n_parking_spots, simulate_days;
     float road_length, speed_limit;
